+++ R/exploreit-package.R
@@ -1,20 +1,23 @@
 #' @keywords internal
 "_PACKAGE"
 
-#' @importFrom stats as.dendrogram as.dist as.hclust cutree kmeans nobs
-#'   model.frame predict
+#' @importFrom stats as.dendrogram as.dist as.hclust cmdscale cor cutree kmeans
+#'   nobs model.frame predict
 #' @importFrom fastcluster hclust
-#' @importFrom tibble as_tibble as_tibble_col
+#' @importFrom tibble tibble as_tibble as_tibble_col
 #' @importFrom data.table as.data.table
-#' @importFrom graphics points symbols
+#' @importFrom graphics lines points symbols
 #' @importFrom utils str
-#' @importFrom chart chart theme_sciviews
-#' @importFrom ggplot2 aes autoplot coord_flip coord_polar element_blank
-#'   element_text expansion geom_hline geom_point geom_segment geom_text ggplot
-#'   scale_x_continuous scale_x_reverse scale_y_continuous scale_y_reverse
-#'   theme
+#' @importFrom chart chart f_aes theme_sciviews
+#' @importFrom ggplot2 aes autoplot coord_fixed coord_flip coord_polar
+#'   element_blank element_text expansion geom_hline geom_point geom_segment
+#'   geom_step geom_text ggplot labs scale_x_continuous scale_x_reverse
+#'   scale_y_continuous scale_y_reverse theme
 #' @importFrom factoextra fviz_nbclust
-#' @importFrom broom augment fix_data_frame tidy
+#' @importFrom broom augment tidy
+#' @importFrom vegan goodness metaMDS monoMDS vegdist wcmdscale
+#' @importFrom MASS isoMDS sammon Shepard
+#' @importFrom ggrepel geom_text_repel
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start