+++ R/exploreit.R
@@ -2,3 +2,8 @@
 #' @export
 #' @seealso [augment.Cluster()], [augment.kmeans()]
 generics::augment
+
+#' @importFrom broom glance
+#' @export
+#' @seealso [glance.mds()]
+broom::glance